+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+### Performance
+- On `MoveGenerator`, Replaced `if` cases on piece attack finding to `switch-case`.
+- On `MoveGenerator`, added `isRacingKings` to store this variant is racing king variant.
+
 ## [1.9.0]
 
 ### Added